Add CUDA JIT to  `calc_gamma_quadratic_stark()` (#126)
* Add basic test for `calc_doppler_width()`

* Refactor `calc_doppler_width()` and add test for vectorized implementation

* Typecast to float

* Add unwrapped cuda implementation of doppler_width

Also typecast all global constants to float

* Add wrapped cuda implementation of calc_doppler_width

* Return cupy array by default

* Add test for `calc_gamma_quadratic_stark()`

* Typecast inputs and remove explicit calls to power operator in favor of repeated multiplication

* Add vectorized implemention of `calc_gamma_quadratic_stark()`

* Add test for vectorized version of `calc_gamma_quadratic_stark()`

* Optimize fourth power operation

* Add CUDA inplementation of `calc_gamma_quadratic_stark()`

* add missed import to test

* fix duplicated tests, move all to opacities_solvers/tests

* add missing __init__.py to tests

* change ion_number to nnot square itself and reorder args in _calc_gamma_quadratic_stark_cuda()

---------
