+++ stardis/radiation_field/opacities/opacities_solvers/tests/test_broadening.py
@@ -7,6 +7,15 @@ from stardis.radiation_field.opacities.opacities_solvers.broadening import (
     calc_doppler_width,
     _calc_doppler_width_cuda,
     calc_doppler_width_cuda,
+    calc_n_effective,
+    _calc_n_effective_cuda,
+    calc_n_effective_cuda,
+    calc_gamma_linear_stark,
+    _calc_gamma_linear_stark_cuda,
+    calc_gamma_linear_stark_cuda,
+    calc_gamma_quadratic_stark,
+    _calc_gamma_quadratic_stark_cuda,
+    calc_gamma_quadratic_stark_cuda,
 )
 
 GPUs_available = cuda.is_available()
@@ -127,4 +136,352 @@ def test_calc_doppler_width_cuda_wrapped_sample_cuda_values(
     assert np.allclose(
         calc_doppler_width_cuda(*map(cp.asarray, arg_list)),
         calc_doppler_width_cuda_wrapped_sample_cuda_values_expected_result,
+    )
+
+
+@pytest.mark.parametrize(
+    "calc_n_effective_sample_values_input_ion_number,calc_n_effective_sample_values_input_ionization_energy,calc_n_effective_sample_values_input_level_energy, calc_n_effective_sample_values_expected_result",
+    [
+        (
+            1.0,
+            RYDBERG_ENERGY,
+            0,
+            1.0,
+        ),
+        (
+            np.array(2 * [1]),
+            np.array(2 * [RYDBERG_ENERGY]),
+            np.array(2 * [0]),
+            np.array(2 * [1.0]),
+        ),
+    ],
+)
+def test_calc_n_effective_sample_values(
+    calc_n_effective_sample_values_input_ion_number,
+    calc_n_effective_sample_values_input_ionization_energy,
+    calc_n_effective_sample_values_input_level_energy,
+    calc_n_effective_sample_values_expected_result,
+):
+    assert np.allclose(
+        calc_n_effective(
+            calc_n_effective_sample_values_input_ion_number,
+            calc_n_effective_sample_values_input_ionization_energy,
+            calc_n_effective_sample_values_input_level_energy,
+        ),
+        calc_n_effective_sample_values_expected_result,
+    )
+
+
+@pytest.mark.skipif(
+    not GPUs_available, reason="No GPU is available to test CUDA function"
+)
+@pytest.mark.parametrize(
+    "calc_n_effective_cuda_unwrapped_sample_values_input_ion_number,calc_n_effective_cuda_unwrapped_sample_values_input_ionization_energy,calc_n_effective_cuda_unwrapped_sample_values_input_level_energy,calc_n_effective_cuda_unwrapped_sample_values_expected_result",
+    [
+        (
+            np.array(2 * [1]),
+            np.array(2 * [RYDBERG_ENERGY]),
+            np.array(2 * [0]),
+            np.array(2 * [1.0]),
+        ),
+    ],
+)
+def test_calc_n_effective_cuda_unwrapped_sample_values(
+    calc_n_effective_cuda_unwrapped_sample_values_input_ion_number,
+    calc_n_effective_cuda_unwrapped_sample_values_input_ionization_energy,
+    calc_n_effective_cuda_unwrapped_sample_values_input_level_energy,
+    calc_n_effective_cuda_unwrapped_sample_values_expected_result,
+):
+    arg_list = (
+        calc_n_effective_cuda_unwrapped_sample_values_input_ion_number,
+        calc_n_effective_cuda_unwrapped_sample_values_input_ionization_energy,
+        calc_n_effective_cuda_unwrapped_sample_values_input_level_energy,
+    )
+
+    arg_list = tuple(map(cp.array, arg_list))
+    result_values = cp.empty_like(arg_list[0])
+
+    nthreads = 256
+    length = len(calc_n_effective_cuda_unwrapped_sample_values_expected_result)
+    nblocks = 1 + (length // nthreads)
+
+    _calc_n_effective_cuda[nblocks, nthreads](result_values, *arg_list)
+
+    assert np.allclose(
+        cp.asnumpy(result_values),
+        calc_n_effective_cuda_unwrapped_sample_values_expected_result,
+    )
+
+
+@pytest.mark.skipif(
+    not GPUs_available, reason="No GPU is available to test CUDA function"
+)
+@pytest.mark.parametrize(
+    "calc_n_effective_cuda_sample_values_input_ion_number, calc_n_effective_cuda_sample_values_input_ionization_energy, calc_n_effective_cuda_sample_values_input_level_energy, calc_n_effective_cuda_wrapped_sample_cuda_values_expected_result",
+    [
+        (
+            np.array(2 * [1]),
+            np.array(2 * [RYDBERG_ENERGY]),
+            np.array(2 * [0]),
+            np.array(2 * [1.0]),
+        ),
+    ],
+)
+def test_calc_n_effective_cuda_wrapped_sample_cuda_values(
+    calc_n_effective_cuda_sample_values_input_ion_number,
+    calc_n_effective_cuda_sample_values_input_ionization_energy,
+    calc_n_effective_cuda_sample_values_input_level_energy,
+    calc_n_effective_cuda_wrapped_sample_cuda_values_expected_result,
+):
+    arg_list = (
+        calc_n_effective_cuda_sample_values_input_ion_number,
+        calc_n_effective_cuda_sample_values_input_ionization_energy,
+        calc_n_effective_cuda_sample_values_input_level_energy,
+    )
+    assert np.allclose(
+        calc_n_effective_cuda(*map(cp.asarray, arg_list)),
+        calc_n_effective_cuda_wrapped_sample_cuda_values_expected_result,
+    )
+
+
+@pytest.mark.parametrize(
+    "calc_gamma_linear_stark_sample_values_input_n_eff_upper,calc_gamma_linear_stark_sample_values_input_n_eff_lower,calc_gamma_linear_stark_sample_values_input_electron_density, calc_gamma_linear_stark_sample_values_expected_result",
+    [
+        (
+            1,
+            0,
+            (0.51 * 0.642) ** (-3 / 2),
+            1.0,
+        ),
+        (
+            np.array(2 * [1]),
+            np.array(2 * [0]),
+            np.array(2 * [(0.51 * 0.642) ** (-3 / 2)]),
+            np.array(2 * [1.0]),
+        ),
+    ],
+)
+def test_calc_gamma_linear_stark_sample_values(
+    calc_gamma_linear_stark_sample_values_input_n_eff_upper,
+    calc_gamma_linear_stark_sample_values_input_n_eff_lower,
+    calc_gamma_linear_stark_sample_values_input_electron_density,
+    calc_gamma_linear_stark_sample_values_expected_result,
+):
+    assert np.allclose(
+        calc_gamma_linear_stark(
+            calc_gamma_linear_stark_sample_values_input_n_eff_upper,
+            calc_gamma_linear_stark_sample_values_input_n_eff_lower,
+            calc_gamma_linear_stark_sample_values_input_electron_density,
+        ),
+        calc_gamma_linear_stark_sample_values_expected_result,
+    )
+
+
+@pytest.mark.skipif(
+    not GPUs_available, reason="No GPU is available to test CUDA function"
+)
+@pytest.mark.parametrize(
+    "calc_gamma_linear_stark_cuda_unwrapped_sample_values_input_n_eff_upper,calc_gamma_linear_stark_cuda_unwrapped_sample_values_input_n_eff_lower,calc_gamma_linear_stark_cuda_unwrapped_sample_values_input_electron_density,calc_gamma_linear_stark_cuda_unwrapped_sample_values_expected_result",
+    [
+        (
+            np.array(2 * [1]),
+            np.array(2 * [0]),
+            np.array(2 * [(0.51 * 0.642) ** (-3 / 2)]),
+            np.array(2 * [1.0]),
+        ),
+    ],
+)
+def test_calc_gamma_linear_stark_cuda_unwrapped_sample_values(
+    calc_gamma_linear_stark_cuda_unwrapped_sample_values_input_n_eff_upper,
+    calc_gamma_linear_stark_cuda_unwrapped_sample_values_input_n_eff_lower,
+    calc_gamma_linear_stark_cuda_unwrapped_sample_values_input_electron_density,
+    calc_gamma_linear_stark_cuda_unwrapped_sample_values_expected_result,
+):
+    arg_list = (
+        calc_gamma_linear_stark_cuda_unwrapped_sample_values_input_n_eff_upper,
+        calc_gamma_linear_stark_cuda_unwrapped_sample_values_input_n_eff_lower,
+        calc_gamma_linear_stark_cuda_unwrapped_sample_values_input_electron_density,
+    )
+
+    arg_list = tuple(map(cp.array, arg_list))
+    result_values = cp.empty_like(arg_list[0], dtype=float)
+
+    nthreads = 256
+    length = len(calc_gamma_linear_stark_cuda_unwrapped_sample_values_expected_result)
+    nblocks = 1 + (length // nthreads)
+
+    _calc_gamma_linear_stark_cuda[nblocks, nthreads](result_values, *arg_list)
+    print(result_values)
+    assert np.allclose(
+        cp.asnumpy(result_values),
+        calc_gamma_linear_stark_cuda_unwrapped_sample_values_expected_result,
+    )
+
+
+@pytest.mark.skipif(
+    not GPUs_available, reason="No GPU is available to test CUDA function"
+)
+@pytest.mark.parametrize(
+    "calc_gamma_linear_stark_cuda_wrapped_sample_values_input_n_eff_upper,calc_gamma_linear_stark_cuda_wrapped_sample_values_input_n_eff_lower,calc_gamma_linear_stark_cuda_wrapped_sample_values_input_electron_density,calc_gamma_linear_stark_cuda_wrapped_sample_values_expected_result",
+    [
+        (
+            np.array(2 * [1]),
+            np.array(2 * [0]),
+            np.array(2 * [(0.51 * 0.642) ** (-3 / 2)]),
+            np.array(2 * [1.0]),
+        ),
+    ],
+)
+def test_calc_doppler_width_cuda_wrapped_sample_cuda_values(
+    calc_gamma_linear_stark_cuda_wrapped_sample_values_input_n_eff_upper,
+    calc_gamma_linear_stark_cuda_wrapped_sample_values_input_n_eff_lower,
+    calc_gamma_linear_stark_cuda_wrapped_sample_values_input_electron_density,
+    calc_gamma_linear_stark_cuda_wrapped_sample_values_expected_result,
+):
+    arg_list = (
+        calc_gamma_linear_stark_cuda_wrapped_sample_values_input_n_eff_upper,
+        calc_gamma_linear_stark_cuda_wrapped_sample_values_input_n_eff_lower,
+        calc_gamma_linear_stark_cuda_wrapped_sample_values_input_electron_density,
+    )
+    assert np.allclose(
+        calc_gamma_linear_stark_cuda(*map(cp.asarray, arg_list)),
+        calc_gamma_linear_stark_cuda_wrapped_sample_values_expected_result,
+    )
+
+
+c4_prefactor = (ELEMENTARY_CHARGE**2 * BOHR_RADIUS**3) / (
+    36.0 * PLANCK_CONSTANT * VACUUM_ELECTRIC_PERMITTIVITY
+)
+
+
+@pytest.mark.parametrize(
+    "calc_gamma_quadratic_stark_sample_values_input_ion_number,calc_gamma_quadratic_stark_sample_values_input_n_eff_upper,calc_gamma_quadratic_stark_sample_values_input_n_eff_lower, calc_gamma_quadratic_stark_sample_values_input_electron_density,  calc_gamma_quadratic_stark_sample_values_input_temperature,calc_gamma_quadratic_stark_sample_values_expected_result",
+    [
+        (
+            1,  # ion_number
+            1.0,  # n_eff_upper
+            0.0,  # n_eff_lower
+            1.0e-19
+            / BOLTZMANN_CONSTANT
+            * (36 * c4_prefactor) ** (-2.0 / 3.0),  # electron_density
+            1.0,  # temperature
+            1,  # Expected output
+        ),
+        (
+            np.array(2 * [1], dtype=int),
+            np.array(2 * [1.0]),
+            np.array(2 * [0.0]),
+            np.array(
+                2 * [1.0e-19 / BOLTZMANN_CONSTANT * (36 * c4_prefactor) ** (-2.0 / 3.0)]
+            ),
+            np.array(2 * [1.0]),
+            np.array(2 * [1.0]),
+        ),
+    ],
+)
+def test_calc_gamma_quadratic_stark_sample_values(
+    calc_gamma_quadratic_stark_sample_values_input_ion_number,
+    calc_gamma_quadratic_stark_sample_values_input_n_eff_upper,
+    calc_gamma_quadratic_stark_sample_values_input_n_eff_lower,
+    calc_gamma_quadratic_stark_sample_values_input_electron_density,
+    calc_gamma_quadratic_stark_sample_values_input_temperature,
+    calc_gamma_quadratic_stark_sample_values_expected_result,
+):
+    assert np.allclose(
+        calc_gamma_quadratic_stark(
+            calc_gamma_quadratic_stark_sample_values_input_ion_number,
+            calc_gamma_quadratic_stark_sample_values_input_n_eff_upper,
+            calc_gamma_quadratic_stark_sample_values_input_n_eff_lower,
+            calc_gamma_quadratic_stark_sample_values_input_electron_density,
+            calc_gamma_quadratic_stark_sample_values_input_temperature,
+        ),
+        calc_gamma_quadratic_stark_sample_values_expected_result,
+    )
+
+
+@pytest.mark.skipif(
+    not GPUs_available, reason="No GPU is available to test CUDA function"
+)
+@pytest.mark.parametrize(
+    "calc_gamma_quadratic_stark_sample_values_input_ion_number,calc_gamma_quadratic_stark_sample_values_input_n_eff_upper,calc_gamma_quadratic_stark_sample_values_input_n_eff_lower, calc_gamma_quadratic_stark_sample_values_input_electron_density,  calc_gamma_quadratic_stark_sample_values_input_temperature,calc_gamma_quadratic_stark_sample_values_expected_result",
+    [
+        (
+            np.array(2 * [1], dtype=int),
+            np.array(2 * [1.0]),
+            np.array(2 * [0.0]),
+            np.array(
+                2 * [1.0e-19 / BOLTZMANN_CONSTANT * (36 * c4_prefactor) ** (-2.0 / 3.0)]
+            ),
+            np.array(2 * [1.0]),
+            np.array(2 * [1.0]),
+        ),
+    ],
+)
+def test_calc_gamma_quadratic_stark_cuda_unwrapped_sample_values(
+    calc_gamma_quadratic_stark_sample_values_input_ion_number,
+    calc_gamma_quadratic_stark_sample_values_input_n_eff_upper,
+    calc_gamma_quadratic_stark_sample_values_input_n_eff_lower,
+    calc_gamma_quadratic_stark_sample_values_input_electron_density,
+    calc_gamma_quadratic_stark_sample_values_input_temperature,
+    calc_gamma_quadratic_stark_sample_values_expected_result,
+):
+    arg_list = (
+        calc_gamma_quadratic_stark_sample_values_input_ion_number,
+        calc_gamma_quadratic_stark_sample_values_input_n_eff_upper,
+        calc_gamma_quadratic_stark_sample_values_input_n_eff_lower,
+        calc_gamma_quadratic_stark_sample_values_input_electron_density,
+        calc_gamma_quadratic_stark_sample_values_input_temperature,
+    )
+
+    arg_list = tuple(map(cp.array, arg_list))
+    result_values = cp.empty_like(arg_list[0])
+
+    nthreads = 256
+    length = len(calc_gamma_quadratic_stark_sample_values_expected_result)
+    nblocks = 1 + (length // nthreads)
+
+    _calc_gamma_quadratic_stark_cuda[nblocks, nthreads](result_values, *arg_list)
+
+    assert np.allclose(
+        cp.asnumpy(result_values),
+        calc_gamma_quadratic_stark_sample_values_expected_result,
+    )
+
+
+@pytest.mark.skipif(
+    not GPUs_available, reason="No GPU is available to test CUDA function"
+)
+@pytest.mark.parametrize(
+    "calc_gamma_quadratic_stark_sample_values_input_ion_number,calc_gamma_quadratic_stark_sample_values_input_n_eff_upper,calc_gamma_quadratic_stark_sample_values_input_n_eff_lower, calc_gamma_quadratic_stark_sample_values_input_electron_density,  calc_gamma_quadratic_stark_sample_values_input_temperature,calc_gamma_quadratic_stark_sample_values_expected_result",
+    [
+        (
+            np.array(2 * [1], dtype=int),
+            np.array(2 * [1.0]),
+            np.array(2 * [0.0]),
+            np.array(
+                2 * [1.0e-19 / BOLTZMANN_CONSTANT * (36 * c4_prefactor) ** (-2.0 / 3.0)]
+            ),
+            np.array(2 * [1.0]),
+            np.array(2 * [1.0]),
+        ),
+    ],
+)
+def test_calc_gamma_quadratic_stark_cuda_wrapped_sample_cuda_values(
+    calc_gamma_quadratic_stark_sample_values_input_ion_number,
+    calc_gamma_quadratic_stark_sample_values_input_n_eff_upper,
+    calc_gamma_quadratic_stark_sample_values_input_n_eff_lower,
+    calc_gamma_quadratic_stark_sample_values_input_electron_density,
+    calc_gamma_quadratic_stark_sample_values_input_temperature,
+    calc_gamma_quadratic_stark_sample_values_expected_result,
+):
+    arg_list = (
+        calc_gamma_quadratic_stark_sample_values_input_ion_number,
+        calc_gamma_quadratic_stark_sample_values_input_n_eff_upper,
+        calc_gamma_quadratic_stark_sample_values_input_n_eff_lower,
+        calc_gamma_quadratic_stark_sample_values_input_electron_density,
+        calc_gamma_quadratic_stark_sample_values_input_temperature,
+    )
+    assert np.allclose(
+        calc_gamma_quadratic_stark_cuda(*map(cp.asarray, arg_list)),
+        calc_gamma_quadratic_stark_sample_values_expected_result,
     )
